feat(mime): Initial cut at better mime detection
Our current content type detection requires the entire file (so that things like JSON.parse() and

DOMParser() work properly). This is a poor solution for larger files and may exceed the available

memory. To address it, the new mime type stack is designed to accept partial files (generally the

first 16KB or less). This is merely the initial stack and tests. It is not hooked into anything yet.