feat(mime): mime stack now replaces content type stack
For all plugins and importable types (servers included). This allows some types to use large files

by not loading all of the content up front.

BREAKING CHANGE: plugins and applications will need to replace the use of IContentTypeMethod with

a mime stack detection, both for file types and server types. This will be updated in the guides

in the following commit, but minimally:

- make a mime detection
- adjust ProviderEntry arguments
- adjust registration of import UIs to match the mime type (if you changed it)
- add directive to ProviderImportUI constructor e.g. new ProviderImport('<theDirective></theDirective>');